Introducing a possibility to define atom kinds:
Here I briefly describe what was done by going through the list of files and
functions which were modified:

params.ini  - now the type of PP can be defined not only with
the number, but you can also provide the name of the atom: "O". It
should also work. I would advise  to use symbols instead of numbers,
because of the following problem:
Let say two define two additional Carbon atoms Csp2, Csp.Then the O atom
is shifted to the position 10, instead of 8. If you don't realize that -
you will use atom 8 for the PP, which now is Csp and not O. While if you
define the probeType as "O" it will always work correctly.

atomtypes.ini - now the file MUST contain  at least 4 columns: Rmin,
epsilon, atom number from the periodic table, symbol (can be Csp2 for
example).

elements.py - this file should not be used any more for any functions
which perform calculations, because all the necessary data are contained
in atomtypes.ini. But this file is used for the functions which perform
plotting, because it contains the information about "atom colours".

pyProbeParticle/common.py: loadSpecies function.
Now it basically contains only one call of np.genfromtxt which makes it
obligatory to have at least 4 columns in atomtypes.ini file

pyProbeParticle/HighLevel.py: computeLJ function
I've removed the call of loadSpecies function, and put it into
generateLJFF.py. In this way it seems to be more logically correct

pyProbeParticle/HighLevel.py: parseAtoms:
Now it depends also on the FFparams (LJ parameters loaded from the
atomtypes.ini file).

And all the function calls were of course modified in generateLJFF.py to
keep consistency.

pyProbeParticle/basUtils.py: loadAtoms. Now this function has only one
argument (filename) and does not depend on the elements.ELEMENT_DICT.
pyProbeParticle/basUtils.py: loadAtomsCube. Now this function has only
one argument (filename) and does not depend on the
elements.ELEMENT_DICT.

pyProbeParticle/basUtils.py: findBonds and getAtomColors. Now both these
functions require FFparams to be provided.

plot_results.py: I made this script to be consistent with the previous
changes. In addition I modify the code in such a way that it is possilbe
now to plot atoms and bonds in PBC.