Added HeightMap and ESMap to AuxMaps, should be complete now. ESMap generation fixed.
Previously ESMap was calculated by interpolating from precalculated electrostatic field. However, the electrostatic field is not saved to a separate buffer anymore, so the ESMap cannot be calculated this way anymore. An easier fix was to just recalculate the electrostatic field during the evaluation of the ESMap, which is how it works now.

Additionally, changed time.clock() to time.time() in fieldOCL because time.clock() got removed in Python 3.8.