New version of generator that combines the AFMulator and AuxMaps.
This version is more minimal in that it modifies less parameters by default. Instead, there are several optional methods for doing the most common augmentations and randomizations, as well as several callback methods that can be used to call the augmentations or otherwise modify parameters at runtime.

Moved eval methods to the top in AFMulator. Replaced BrCl2Benzene example xyz with one that contains actually calculated values for atom charges.