Changed scan_window definition in AFMulator.
The scan window in AFMulator is now fixed by the user and the positions of the atoms are taken as they are given instead of shifting them to the center of the force field cell. This allows for more flexibility, but requires the user to make sure that the molecule is in the correct position and that the scan window is positioned correctly with respect to the force field cell. In particular, this allows putting the molecule to the bottom of the simulation cell so that the force field cell can be made much smaller, making the time spent on force field calculation smaller by a significant margin.

The molecule position handling in InverseAFMtrainer is changed to reflect the changes in AFMulator.