Big changes to CorrectionTrainer, some to Mutator
CorrectionTrainer is now a class that inherits InverseAFMtrainer. The __next__ function returns two AFM images and
the two corresponding xyz representations in batches. The size of batches is specified by the user in initialization.
The logic of generatePair has been changed: now the function only does the mutation and returns the original and mutant molecules
and the AFM images are taken in __next__.

Mutator can now do a series of mutations up to user-specified maxMutations argument.