Hartree potentials now work with the Afmulator + other fixes
Added kernel that computes the LJ force and adds to it the field from a Hartree potential. The grid is hard-coded to be rectangular for now.

Fixed ForceField_LJC.downloadFF to use proper memory layout for the returned force-field array. Cleaned up a bit ForceField_LJC.makeFF by deleting unused code lines and arguments.

The Afmulator was not taking the origin of the force-field box into account during the scan. Fixed by subtracting the origin from the grid position before creating the scan grid.