Various fixes and clean-up to AFMulator.
The convolution kernel was leaving the last step of the relax scan unused. Changed the convolution kernel so that is goes one step further. This makes the resulting AFM images one step larger in z-direction.

Removed the amplitude argument from AFMulator initialization. The convolution kernel always uses the step size of the scanner so the amplitude argument along with the number of steps was redundant. The amplitude is now instead determined as the number of steps times the scanner step size.

Moved the tipR0, tipStiffness, and npbc attributes of the AFMulator class to the init arguments so that they can be easily changed when creating the class object. Also changed the default tip distance to 3.0, since it should be more standard.

Updated the docstring in AFMulator to reflect the changes and to be more detailed.