Fixed output units in AFMulator + small optimization.
The output units in AFMulator were not taking into account the cantilever parameters and a unit conversion factor was missing as well. Added the cantilever resonance frequency and stiffness as adjustable parameters to the AFMulator class init.

The Hartree potential gradient calculation now checks if the potential grid already matches the force field grid, and if it does then it uses a simpler kernel that does not need to use interpolation.