Hartree potentials now work with non-rectangular grids.
While testing this with a potential from VASP, found that the output was garbage because the potential was not very smooth. Worked fine with a potential from FHI-aims on the same system. This probably stems from the use of the finite difference derivative with the point charges on the tip. Probably would work better if the tip used the proper spherical harmonic density.

Also changed the tip relaxation interpolation to use periodic boundaries instead of reflective.

Added a check into AFMulator that issues a warning if the scan window extends over a non-periodic boundary in the force-field grid.