Added some utilities for ease of use and a small fix to OCL code.
The AFMulator now computes the lvec automatically if not given. The OCL environment is automatically initialized if not already initialized instead of throwing an error.

Added a function alongside the AFMulator for quickly doing a simulation for a given file path.

Added a functions to oclUtils for quickly quering available platforms and quickly initializing the environment for the force field and the scanner.

Fixed signature in OCL interpolation functions so that it also compiles on Intel iGPU ICD platform.