Redesigning the GUI. Mostly functional.
Changed the scan parameters from start and end specification to specifying the scan region size and location. The scan start in z is now specified by a distance relative to the molecule, so that there is no need to deal with absolute coordinates, which vary a lot between system.

Sometimes crashes when changing the scan size or step.

The amplitude now determines the total scan size in z direction. This makes it simpler so that it's not possible to input scan size that is smaller than the amplitude, but it also breaks the df curve plotting.

Added an file open dialog, so that the input file does not have to be in the directory.

Fixed the image and data saving buttons.