Fixed GUI crash when changing the scan window size.
The force field buffers were not being reallocated even when the size of the force field grid increased.

Recalculating the force field on a different grid using the Hartree potentials is presenting a problem, though. The reikna FFT package recompiles the FFT operation for different grid sizes which takes considerable amount of time, which makes the GUI very laggy when changing the scan size.

Added a command line option for verbosity level.