Added GUI option for using periodic boundary condition + small optimization.
Made plotting it a bit faster by changing the plot update to only set new data to the plot when the size of the plot did not change. The plotting is still the biggest bottleneck in the application. Probably matplotlib is just not the correct tool to use for fast interactive plotting.