Deleted old OCL related functions no longer in use (#37)
The deleted functions were either used nowhere or in one of the RunOpenCLbatch*.py scripts or GUImorse.py, which we were discussing should be deleted anyway.

`ocl.HighLevel.py did not have any functions in use except for genFFSampling and loadSpecies, which were already defined in common.py, so I deleted the whole module. I just had to modify `common.loadSpecies` a bit so that it checks if the file exists.