C++ code compilation for pip wheel (#92)
Add setup.py and MANIFEST.in in order to include the compiled C++ extensions into the pip wheel. The C++ source files are modified to be compatible with the Microsoft C++ compiler, so that the pip wheel can also be built on Windows.

In order not to make the C++ extension recompile every time by default, add an environment variable PPAFM_RECOMPILE that needs to be set to a non-empty value if you want the code recompile.

Also some reorganization to move non-essential Python modules and C++ source files to a new submodule called dev inside ppafm. This module is excluded from the pip wheel.