Fix for sign issues in electrostatics and Pauli calculations. (#124)
The sign of the Pauli force was coming out wrong in the pyridine density overlap example. The underlying issue was that the FFT convolution for calculating the force had one minus sign missing. This also affects the electrostatics calculation that had another minus sign missing due to wrong input units, which was previously cancelling out.

Also changed the Pauli prefactor in the pyridine example from 1.0 to 18.0, because that is the default and is also used in a publication.