Parameter saving/loading from params.ini for AFMulator and the GUI (#139)
Adds methods for constructing or loading an AFMulator instance from a params.ini file and saving the current settings to a params.ini file. Also adds the corresponding bindings to the GUI script and buttons to the GUI for accessing this feature.

This should help in bridging the gap between the usage of the CPU scripts and the OpenCL API/GUI.

This required adding the entries FFgrid{0,A,B,C} to the parameter dict to denote the force field grid origin and lattice vectors, because the AFMulator handles the force field lattice separate from the sample lattice vectors. I hope that this does not cause any confusion since these are not used by the CPU scripts.