Removed locale change in ReadNumsUpTo that was causing trouble. (#141)
Manipulating the locale inside the C++ data reading function was causing subsequent regular file reads in Python to fail with a UnicodeDecodeError. This deletes the offending lines. It seems this was originally a fix for some incompatible interpretation of decimal separators (, vs. .), but this could not replicated for the moment.