Fix to OpenCL compilation paths. (#151)
The compilation of the OpenCL code was not taking into account that there may be spaces in the source path. Thanks to @alexriss for pointing this out in https://github.com/Probe-Particle/ppafm/issues/110#issuecomment-1516538754.

There is one edge case that needed to be taken into account, namely that the OpenCL platform that we use for the automated tests on Github, `pocl`, doesn't handle quotes or spaces in the file path properly. This is actually fixed in the latest versions (https://github.com/pocl/pocl/issues/962), but the version in Ubuntu 22.04 doesn't seem to have this fix.

I also took the opportunity to switch from using `os.path` to using `pathlib.Path` for the paths, which should be more general.