Implemented DFT-D3 in the CPU code. (#166)
I went ahead and also implemented the DFT-D3 calculation in the CPU code now that it's still fresh in my mind. I added a new script generateDFTD3.py for accessing this feature. It replaces the old LJ script in the pyridine example run script. Also added a test to make sure that the CPU and GPU versions return the same result for the force field.