Set fixed locale during array read. (#168)
The locale change was previously removed because it was causing trouble with file reading in Python. However, the removal caused the array read in C++ to fail when using the GUI, because pyqt seems to set a different environment locale, which on some systems uses , (comma) as a decimal separator. This sets a fixed locale temporarily during the array read and then sets it back so that everything should work as intended.