CI wheels build (#178)
Add a job to the PyPI CI workflow that builds wheels for all platforms (64-bit).

The job generates a separate wheel for every Python version >=3.7. I don't think there is actually any difference in the wheels between the different Python versions, but I didn't find a way to just generate one. I tested that the wheel build works, but I didn't test how it interacts with the upload to PyPI.

The minimum Python version had to be bumped up to 3.7 because there does not exist a new-enough version of setuptools for 3.6.

Removed the lookup for MSVC variables on Windows because there is a Github action that does it better.