Overhaul of CLI arguments (#185)
Adds to ppafm.common a new class CLIParser, which holds a list of common commandline arguments that can be added to the scripts.

The arguments are divided into two categories: those which directly correspond to one of the parameters in the ppafm.common.params dictionary, and others. The parameter arguments are checked so that their names match with the names in the parameter dictionary to avoid problems.

All the main CLI scripts are changed to use the new parser. But not all arguments are transferred to the pre-defined list; only the ones which are used in more than one place.

A couple of the arguments are renamed:
- data_format -> output_format to contrast with the added new argument input_format.
- Bpower-> Bpauli in order to be consistent with the parameter name in the params dictionary.
- densityMayBeNegative -> no_negative_check in conv_rho.py for a more descriptive name.

---------

Co-authored-by: Martin Ondracek <ondracek@fzu.cz>