Implement `PpafmParameters` class based on the `pydantic.BaseModel` (#275)
Previously, the run parameters were stored in a global dictionary called `params`. This
solution was sub-optimal, as it was very tricky to do multiple runs to avoid "contamination"
of parameters by the previous run. In this PR I introduce the parameters class `PpafmParameters`.
Now, every run will have its set of parameters. Moving forward, we can simplify the signature of
many functions relying on the parameters provided by the `PpafmParameters` object. The
original `params` dictionary is now completely removed.

Previously, the default values for parameters were scattered around. Now, all the default values are
set by the `PpafmParameters`, which allows the removal of the `params.ini` file which was part
of the package distribution.

The infrastructure built here offers an easy way to  switch input format from `ini`
to `toml`, as suggested in the issue #153. To allow easy switching `to_file` method is implemented.
It allows dumping the current parameters to the `toml` format.

Finally, tests were added to make sure the `PpafmParameters`  class behaves as expected.