Fix API docs (#305)
The API docs were blank because the new dependencies from #275 were missing in the ReadTheDocs build.

The badge on the front page was not showing the failure. This is because the failed imports are treated as warnings instead of errors. So I also added fail_on_warning: true to the config and fixed all of the remaining warnings so that there is at least some indication that the build is failing if this happens again in the future.