Fix spline interpolation segfaults. (#309)
Fixes #308
Fixes #279

This PR moves the spline parameters in the TIP namespace into a struct in the C++ code and introduces a corresponding Python class for managing the parameter arrays. A pointer to the parameters are now explicitly passed into the relevant functions instead of being set on a global level. This gets rid of the problem where the parameter array pointer could still be set even after the arrays were freed.

Also fixes one uninitialized variable in the spline interpolation function that was sometimes segfaulting.

All tests are now enabled and seem to be passing without problems again.